.tools/requirements.txt
# Python 3.11
pyyaml==6.0.2
nbformat==5.10.4